+++ src/gui.py
@@ -28,6 +28,8 @@ class Ctrl:
       listener.extraUtils(self.model) #ensure the listener can send changes to the model
 
       controller.add_listener(listener)
+
+
       self.root.mainloop()
       print "Press Enter to quit..."
       #need to replace this block with some input from View to trigger exit and so on
@@ -54,7 +56,7 @@ class Model():
       return self.txt
 
    def textChanged(self, text):
-      if len(self.txt) < 12:
+      if len(self.txt) < 26:
          self.txt = self.txt + text
       else:
          self.txt = text
